refactor: merge to onefile
